move quiz commmand handler to new file
